fix(warnings): eliminate all engine compiler warnings from clean build
- ReplayPlayer.cpp:147: explicit static_cast<double>(m_playback_speed)
  to fix -Wdouble-promotion (float * 60.0 was silently widening)
- simulation_demo/main.cpp:121-126: mark sides/radii/r_cols/g_cols/b_cols
  [[maybe_unused]] with //Reserved comments -- visual rendering deferred
- multiplayer_demo/main.cpp, scripted_demo/main.cpp: [[maybe_unused]] on argc/argv
- SteamIntegration.cpp:83: add (void)dialog in #else branch to suppress
  -Wunused-parameter when STEAMWORKS_SDK is not defined

Result: cmake --build build produces 0 engine warnings (only Catch2 third-party remains)

+++ Engine/Replay/src/ReplayPlayer.cpp
@@ -144,7 +144,7 @@ namespace Shape::Replay {
 
     void ReplayPlayer::update(double real_delta) {
         if (m_state != PlayerState::Playing) return;
-        const double ticks_per_sec = 60.0 * m_playback_speed;
+        const double ticks_per_sec = 60.0 * static_cast<double>(m_playback_speed);
         const double advance = real_delta * ticks_per_sec;
         const uint64_t ticks_to_advance = static_cast<uint64_t>(advance);
         for (uint64_t i = 0; i < ticks_to_advance; ++i) {
